ensure tests pass
