+++ tests/create_meta/model_card.md
@@ -1,0 +1,18 @@
+---
+model-index:
+- name: all-MiniLM-L6-v2
+  results:
+  - dataset:
+      config: default
+      name: MTEB BornholmBitextMining
+      revision: 3bc5cfb4ec514264fe2db5615fac9016f7251552
+      split: test
+      type: strombergnlp/bornholmsk_parallel
+    metric:
+      type: f1
+      value: 0.2968132161955691
+    task:
+      type: BitextMining
+tags:
+- mteb
+---
