+++ tests/test_cli.py
@@ -32,3 +32,25 @@ def test_run_task(
     assert f"{task_name}.json" in [
         f.name for f in list(results_path.glob("*.json"))
     ], f"{task_name} not found in output folder"
+
+
+def test_create_meta():
+    test_folder = Path(__file__).parent
+    output_folder = test_folder / "create_meta"
+    results = (
+        output_folder / "all-MiniLM-L6-v2" / "8b3219a92973c328a8e22fadcfa821b5dc75636a"
+    )
+    output_path = output_folder / "model_card.md"
+    command = f"mteb create_meta --results_folder {results} --output_path {output_path} --overwrite"
+    result = subprocess.run(command, shell=True, capture_output=True, text=True)
+    assert result.returncode == 0, "Command failed"
+
+    assert output_path.exists(), "Output file not created"
+
+    with output_path.open("r") as f:
+        meta = f.read()
+
+    with (output_folder / "model_card_gold.md").open("r") as f:
+        gold = f.read()
+
+    assert meta == gold, "Output does not match gold"
